+++ composeApp/src/androidMain/kotlin/org/example/project/presentation/feature/qr_code_menu/screens/qr_code_screen/InfrastructurePrinterVkpAPI.kt
@@ -4,16 +4,16 @@ package org.example.project.presentation.feature.qr_code_menu.screens.qr_code_sc
 interface InfrastructurePrinterVkpAPI {
     fun printOnVKP(description: String?,
                    textBarcode: String,
-                   heightQRCodeMM:Int,
+                   heightQRCodeMM:Float,
                    fontSize:Float)
     fun connectUSB()
     fun <T>getQRCode(content: String,
-                     heightMm: Int
+                     heightMm: Float
     ):T
 
     fun <T>getTitleProduct(
         content: String,
-        fontSize: Int
+        fontSize: Float
     ):T
 
 }
